Flushed out the main page, seems to work for different nutrition types, lots of testing still needed. Major Todo: Styling (notice the 'Select Nutrient' button is at the top of the page... it's 5am so I will not be fixing that but hopefully it's not too bad), plus potentially more rigorous testing on the results. You'll notice in the logs of that there's some fishy indexing going on in some cases

+++ server/services/kroger.py
@@ -83,7 +83,7 @@ def Kroger_parse_product(product):
     product_info['Name'] = product['description']
     product_info['Unit Price'] = product['items'][0]['price']['regular']
     size = (product['items'][0]['size']).split(" ")
-    product_info['Unit Size'] = int(size[0])
+    product_info['Unit Size'] = float(size[0])
     product_info['Unit Size units'] = size[1]
     product_info['Category'] = product['categories'][0]
     product_info['upcID'] = product['upc']
